backend/requirements.txt
# AI Resume Screener — Python Dependencies
# =========================================
# Install with:  pip install -r requirements.txt

# ── Web Framework ──────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20        # Required for UploadFile (file uploads)

# ── Data Validation ────────────────────────────────────────────
pydantic==2.10.4
pydantic-settings==2.7.1

# ── AI / Embeddings ────────────────────────────────────────────
sentence-transformers==3.3.1    # all-MiniLM-L6-v2 model
faiss-cpu==1.9.0.post1          # Vector similarity search

# ── Resume Parsing ─────────────────────────────────────────────
pymupdf==1.25.1                 # PDF text extraction  (import as fitz)
python-docx==1.1.2              # DOCX text extraction (import as docx)

# ── Numerical ──────────────────────────────────────────────────
numpy==2.2.1
scikit-learn==1.6.0             # Optional: for extra metrics / preprocessing
